requirements.txt
typing-extensions
langchain-openai
langchain-experimental
pandas
yfinance
praw
feedparser
stockstats
eodhd
langgraph
chromadb
setuptools
backtrader
akshare
tushare
finnhub-python
parsel
requests
tqdm
pytz
redis
chainlit
rich
questionary
python-dotenv  # Load environment variables from .env file
langchain_anthropic
langchain-google-genai
scipy  # Required for options Greeks calculation (Black-Scholes model)
numpy  # Required for options calculations
torch  # Required for RL model (DQN)
matplotlib  # Required for training visualization
streamlit  # Web interface for interactive dashboard
plotly  # Interactive financial charts
kaleido  # Static image export for plotly
